+++ tests/poll/test_views.py
@@ -12,10 +12,11 @@ from bulbs.content.models import Content
 from bulbs.poll.models import Poll, Answer
 from bulbs.utils.test import (
     random_title,
-    mock_vault,
     BaseAPITestCase,
 )
 from bulbs.utils.test.make_vcr import make_vcr
+from bulbs.utils.test.mock_vault import mock_vault
+
 from .common import SECRETS
 
 vcr = make_vcr(__file__)  # Define vcr file path
@@ -235,9 +236,9 @@ class AnswerAPITestCase(BaseAPITestCase):
     @mock_vault(SECRETS)
     def test_router_registered(self):
         list_url = reverse('answer-list')
-        self.assertEqual(list_url, '/api/v1/answer/')
+        self.assertEqual(list_url, '/api/v1/poll-answer/')
         detail_url = reverse('answer-detail', kwargs={'pk': 1})
-        self.assertEqual(detail_url, '/api/v1/answer/1/')
+        self.assertEqual(detail_url, '/api/v1/poll-answer/1/')
 
     @vcr.use_cassette()
     @mock_vault(SECRETS)