+++ CHANGELOG.md
@@ -1,5 +1,19 @@
 # django-bulbs Change Log
 
+## Development
+
+- `utils.test.BaseIndexableTestCase` now waits for ES shard startup, to reduce flaky tests querying ES before it's ready.
+
+## Version 0.8.5
+
+- Added poll image field to polls
+- Added images and image url fields to poll answers
+
+## Version 0.8.4
+
+- Started `videos` app
+- Added `SeriesDetailView` to `videos/views`
+
 ## Version 0.8.3
 
 - Reduce RSS view caching from 10 to 5 min (per social squad)